chData 관련 코드 channelId 타입 수정

+++ retriever/src/main/java/com/team7/retriever/service/ChDataService.java
@@ -33,7 +33,7 @@ public class ChDataService {
     }
 
     // 채널 아이디로 조회
-    public List<ChData> getChannelDataByChannelId(int channelId) {
+    public List<ChData> getChannelDataByChannelId(long channelId) {
         return chDataRepository.findByChannelId(channelId);
     }
